pyproject.toml
[tool.poetry]
name = "fred-visualization-app"
version = "0.1.0"
description = "A Streamlit app for visualizing FRED, ALFRED, and Maps data"
authors = ["Alex Jensen <thanotosq@gmail.com>"]
packages = [{include = "fred-visualization-app"},]

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<3.12"
pandas = "^2.0.0"
streamlit = "1.22.0"
fredapi = "0.5.1"
python-dotenv = "1.0.0"
jinja2 = "3.1.2"
numpy = "1.23.5"
scipy = "1.10.1"
tenacity = "8.2.2"
tornado = "6.2"
plotly = "^5.22.0"
dash = "^2.17.1"
werkzeug = "~2.3.7"
flask = "~2.3.3"
dash-bootstrap-components = "^1.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pytest-mock = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"